Create debathena-metrics package for gathering information on how Athena is being used.
This package will start	a daemon at the	beginning of a login
session. This daemon captures any path executed	  by the user over the
course of their session, any packages installed	  by the user, and the
total length of	the login session.

The collected information is sent to wslogger when the user is logged
out.